+++ Sources/LocalMindApp/LocalMindApp.swift
@@ -1,3 +1,4 @@
+import AppKit
 import LocalMindCore
 import LocalMindLicense
 import LocalMindRAG
@@ -6,9 +7,16 @@ import SwiftUI
 
 @main
 struct LocalMindApp: App {
+    init() {
+        NSApplication.shared.setActivationPolicy(.regular)
+    }
+
     var body: some Scene {
         WindowGroup {
             ContentView()
+                .onAppear {
+                    NSApp.activate(ignoringOtherApps: true)
+                }
         }
     }
 }